coll/han: Add SMSC-based alltoall to HAN
Add Alltoall algorithm to coll/han.  Each rank on one host is assigned
a single partner on a remote host and vice versa.  Then the rank
collects all the data its partner will need to receive from it's host,
and sends it in one large send, and likewise receives it's data in one
large recv, then cycles to the next host.

This algorithm is only selected when SMSC component has ability to
direct-map peer memory, which only exists for XPMEM module.

Signed-off-by: Luke Robison <lrbison@amazon.com>